S3: list_object_versions() now returns the correct NextToken when paginating
